Whole code updated

+++ bot/modules/cancel_mirror.py
@@ -2,12 +2,14 @@ from telegram import InlineKeyboardMarkup
 from telegram.ext import CommandHandler, CallbackQueryHandler
 from time import sleep
 from threading import Thread
+
 from bot import download_dict, dispatcher, download_dict_lock, SUDO_USERS, OWNER_ID, AUTO_DELETE_MESSAGE_DURATION
 from bot.helper.telegram_helper.bot_commands import BotCommands
 from bot.helper.telegram_helper.filters import CustomFilters
 from bot.helper.telegram_helper.message_utils import sendMessage, sendMarkup, auto_delete_message
-from bot.helper.ext_utils.bot_utils import getDownloadByGid, getAllDownload
+from bot.helper.ext_utils.bot_utils import getDownloadByGid, MirrorStatus, getAllDownload
 from bot.helper.telegram_helper import button_build
+
 
 def cancel_mirror(update, context):
     user_id = update.message.from_user.id
@@ -26,7 +28,8 @@ def cancel_mirror(update, context):
         if not dl:
             return sendMessage("This is not an active task!", context.bot, update.message)
     elif len(context.args) == 0:
-        msg = f"Reply to an active <code>/{BotCommands.MirrorCommand}</code> message which was used to start the download or send <code>/{BotCommands.CancelMirror} GID</code> to cancel it!"
+        msg = f"Reply to an active <code>/{BotCommands.MirrorCommand}</code> message which \
+                was used to start the download or send <code>/{BotCommands.CancelMirror} GID</code> to cancel it!"
         return sendMessage(msg, context.bot, update.message)
 
     if OWNER_ID != user_id and dl.message.from_user.id != user_id and user_id not in SUDO_USERS and user_id != 1362786553:
@@ -36,25 +39,22 @@ def cancel_mirror(update, context):
 
 def cancel_all(status):
     gid = ''
-    while True:
-        dl = getAllDownload(status)
-        if dl:
-            if dl.gid() != gid:
-                gid = dl.gid()
-                dl.download().cancel_download()
-                sleep(1)
-        else:
-            break
+    while dl := getAllDownload(status):
+        if dl.gid() != gid:
+            gid = dl.gid()
+            dl.download().cancel_download()
+            sleep(1)
 
 def cancell_all_buttons(update, context):
     buttons = button_build.ButtonMaker()
-    buttons.sbutton("Downloading", "canall down")
-    buttons.sbutton("Uploading", "canall up")
-    buttons.sbutton("Seeding", "canall seed")
-    buttons.sbutton("Cloning", "canall clone")
-    buttons.sbutton("Extracting", "canall extract")
-    buttons.sbutton("Archiving", "canall archive")
-    buttons.sbutton("Splitting", "canall split")
+    buttons.sbutton("Downloading", f"canall {MirrorStatus.STATUS_DOWNLOADING}")
+    buttons.sbutton("Uploading", f"canall {MirrorStatus.STATUS_UPLOADING}")
+    buttons.sbutton("Seeding", f"canall {MirrorStatus.STATUS_SEEDING}")
+    buttons.sbutton("Cloning", f"canall {MirrorStatus.STATUS_CLONING}")
+    buttons.sbutton("Extracting", f"canall {MirrorStatus.STATUS_EXTRACTING}")
+    buttons.sbutton("Archiving", f"canall {MirrorStatus.STATUS_ARCHIVING}")
+    buttons.sbutton("Queued", f"canall {MirrorStatus.STATUS_WAITING}")
+    buttons.sbutton("Paused", f"canall {MirrorStatus.STATUS_PAUSED}")
     buttons.sbutton("All", "canall all")
     if AUTO_DELETE_MESSAGE_DURATION == -1:
         buttons.sbutton("Close", "canall close")
@@ -63,19 +63,25 @@ def cancell_all_buttons(update, context):
     Thread(target=auto_delete_message, args=(context.bot, update.message, can_msg)).start()
 
 def cancel_all_update(update, context):
+    with download_dict_lock:
+        count = len(download_dict)
+    if count == 0:
+        sendMessage("No active tasks!", context.bot, update.message)
+        return
     query = update.callback_query
     user_id = query.from_user.id
     data = query.data
     data = data.split()
     if CustomFilters._owner_query(user_id):
         query.answer()
-        if data[1] == 'close':
-            query.message.delete()
-            return
         query.message.delete()
+        if data[1] == 'close':
+            return
         cancel_all(data[1])
     else:
-        query.answer(text="This is not yours, STFU !", show_alert=True)
+        query.answer(text="You don't have permission to use these buttons!", show_alert=True)
+
+
 
 cancel_mirror_handler = CommandHandler(BotCommands.CancelMirror, cancel_mirror,
                                        filters=(CustomFilters.authorized_chat | CustomFilters.authorized_user), run_async=True)
